update: create stage processor
